+++ webapp/web/templates/base/base.html
@@ -24,12 +24,12 @@
     <input id="sidebar-toggle" type="checkbox" class="hidden peer" />
 
     <!-- Sidebar -->
-    <div id="sidebar" class="w-64 bg-base-200 shadow-md h-full z-10 peer-checked:hidden">
+    <div id="sidebar" class="sidebar fixed top-0 left-0 w-64 h-full bg-base-200 shadow-md z-20 peer-checked:translate-x-[-100%] peer-checked:opacity-0">
       {% include "base/sidebar.html" %}
     </div>
 
     <!-- Main Content -->
-    <div class="flex-1 flex flex-col">
+    <div id="main-content" class="ml-64 transition-all duration-500 flex-1 flex flex-col peer-checked:ml-0">
       <!-- Navbar -->
       <div class="z-1 w-full bg-base-100">
         {% include "base/navbar.html" %}
