+++ webapp/web/templates/base/navbar.html
@@ -38,7 +38,7 @@
 
   <!-- User Profile -->
   <div class="flex-none">
-    <a href="/users/profile" class="flex items-center gap-2">
+    <a href="/profile/" class="flex items-center gap-2">
       <div class="w-10 h-10 bg-primary rounded-full flex items-center justify-center text-primary-content font-bold">
         <span class="text-lg font-bold">{{ current_user.username[0]|upper }}</span>
       </div>
