+++ webapp/web/templates/profile/form-edit-profile.html
@@ -1,7 +1,7 @@
 <h2 class="text-lg font-bold mb-4">
   Edit Profile
 </h2>
-<form method="POST" action="{{ url_for('users.load_edit_profile') }}">
+<form method="POST" action="{{ url_for('profile.load_edit_profile') }}">
 <div class="mb-2">
   <label class="block text-sm font-medium">
       Full Name
