+++ webapp/web/templates/profile/profile.html
@@ -61,7 +61,7 @@
         </div>
       </div>
       <button class="btn btn-primary mt-4 w-32" 
-        hx-get="/users/load-edit-profile" 
+        hx-get="/profile/load-edit-profile" 
         hx-target="#modal-content" 
         hx-swap="innerHTML" 
         hx-on="htmx:afterRequest: document.getElementById('modal').classList.add('modal-open')">
