+++ webapp/web/templates/role-management/add-role.html
@@ -1,5 +1,5 @@
 <h2 class="text-lg font-bold mb-4">Add New Role</h2>
-<form method="POST" action="{{ url_for('users.add_role') }}">
+<form method="POST" action="{{ url_for('role_management.add_role') }}">
   <div class="mb-4">
     <label class="block text-sm font-medium">Role Name</label>
     <input type="text" name="name" class="input input-bordered w-full" placeholder="Enter role name" required>
