+++ webapp/web/templates/users-management/form-edit-user-role.html
@@ -1,4 +1,4 @@
-<form action="{{ url_for('users.load_edit_user_role', user_id=user.id) }}" method="POST">
+<form action="{{ url_for('users_management.load_edit_user_role', user_id=user.id) }}" method="POST">
     <div class="mb-2">
       <label class="block text-sm font-medium">Full Name</label>
       <input type="text" name="username" class="input input-bordered w-full" value="{{ form.username.data }}" readonly>
