+++ webapp/web/templates/users-management/users-management.html
@@ -89,7 +89,7 @@
               <td class="px-4 py-3">
                 <div class="flex gap-2">
                   <button class="text-blue-500 hover:text-blue-700 transition-colors"
-                  hx-get="{{ url_for('users.load_edit_user_role', user_id=user.id) }}" 
+                  hx-get="{{ url_for('users_management.load_edit_user_role', user_id=user.id) }}" 
                     hx-target="#modal-content" 
                     hx-swap="innerHTML" 
                     hx-on="htmx:afterRequest: document.getElementById('modal').classList.add('modal-open')">
